Conflicts:
	Makefile.am

+++ job.c
@@ -143,8 +143,8 @@ job_write_callback(unused struct bufferevent *bufev, void *data)
 	struct job	*job = data;
 	size_t		 len = EVBUFFER_LENGTH(EVBUFFER_OUTPUT(job->event));
 
-	log_debug("job write %p: %s, pid %ld, output left %lu", job, job->cmd,
-		    (long) job->pid, (unsigned long) len);
+	log_debug("job write %p: %s, pid %ld, output left %zu", job, job->cmd,
+		    (long) job->pid, len);
 
 	if (len == 0) {
 		shutdown(job->fd, SHUT_WR);
